Add prepareExternalResources module
