+++ src/util/prepareResources.ts
